Add serversTransports for proxyProtocol
Though this is not a pretty implementation at all since it cannot be enhanced with other options, this helps TCP services which use the PROXY protocol (https://www.haproxy.org/download/2.0/doc/proxy-protocol.txt) to retrieve the client's IP address. We cannot use `X-Forwarded-For` as it is a HTTP header field.

See:
- https://doc.traefik.io/traefik/reference/routing-configuration/tcp/serverstransport/
- https://github.com/ergochat/ergo/blob/master/docs/MANUAL.md#reverse-proxies

Signed-off-by: Suguru Hirahara <did:key:z6MkvVZk1A3KBApWJXv2Ju4H14ErDfRGxh8zxdXSZ4vACDg5>